requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
ultralytics==8.0.196  # YOLOv8
opencv-python==4.8.1.78
numpy==1.24.3
openai==1.3.0  # GPT API
python-multipart==0.0.6 